feat(quanticstci): adopt 0-indexed grid indices (quanticsgrids 0.2.0) (#613)
Follows the quanticsgrids-rs 0.2.0 migration (tensor4all/quanticsgrids-rs#2):
the workspace now has exactly one indexing convention (0-indexed), and the
five +/-1 conversions at the TCI boundary are gone.

- Bump the quanticsgrids pin to 8214b72 (0.2.0).
- quanticscrossinterpolate_discrete callbacks receive 0-indexed grid
  indices as &[usize] instead of 1-indexed &[i64].
- QuanticsTensorCI2::evaluate takes &[usize] (0-indexed).
- initial_pivots are 0-indexed Vec<Vec<usize>>; cachedata() keys are
  0-indexed quantics indices.
- Removed the (q +/- 1) conversions in quantics_tci.rs and the batched
  path; the TTCache/treetci boundary is now convention-free.
- docs/book/src/conventions.md drops the 1-indexed exception clause;
  quanticstci module/README docs updated; CHANGELOG porting note added.
- tutorial-code and interpolativeqtt/treetci tests updated to the
  0-indexed API.

Closes tensor4all/tensor4all-rs#582.

+++ crates/tensor4all-quanticstci/src/batched/mod.rs
@@ -169,7 +169,7 @@ where
 /// * `output_dims` - Shape of the function output (e.g., `&[3]` for 3-vector,
 ///
 ///   `&[2, 2]` for 2x2 matrix)
-/// * `initial_pivots` - Initial pivot grid indices (optional)
+/// * `initial_pivots` - Initial pivot grid indices (0-indexed, optional)
 /// * `options` - TCI options
 ///
 /// # Returns
@@ -216,7 +216,7 @@ pub fn quanticscrossinterpolate_batched<V, F>(
     grid: &DiscretizedGrid,
     f: F,
     output_dims: &[usize],
-    initial_pivots: Option<Vec<Vec<i64>>>,
+    initial_pivots: Option<Vec<Vec<usize>>>,
     options: QtciOptions,
 ) -> QtciResult<(QuanticsTensorCI2Batched<V>, Vec<usize>, Vec<f64>)>
 where
